Added ability to add and draw tail blocks
Every update of the game, I add the current snake pos
to the start of the tail, I move the snake to the next
block, I check if the snake landed on the food, and if
it does I add another block to the tail. Then I pop off
the last tail block before the paint.

I also added another method to check if the current snake
position collided on any piece of the tail